refactor(onboard): isolate smoke test harness
